Adds support for httpApi
This patch adds support for the new httpApi that aws offers
for reducing cost and latency in serverless applications. It
has been supported for a while in serverless, this just adds
support here to match.

The approach is pretty straightforward so far and may not include
all of the functionality that httpApi permits so far. Those will
come in later patches to this pull request.